Modularized Sorting Hat.
