+++ public/main.js
@@ -1,35 +1,11 @@
 import '../styles/main.scss'; // You have to import your styles for them to work. Comment in this line
-
-const students = [];
-const voldysArmy = []; // starts as an empty array
-
-const houses = [
-  {
-    house: 'gryffindor',
-    crest:
-      'https://static.wikia.nocookie.net/pottermore/images/1/16/Gryffindor_crest.png'
-  },
-  {
-    house: 'slytherin',
-    crest:
-      'https://static.wikia.nocookie.net/pottermore/images/4/45/Slytherin_Crest.png'
-  },
-  {
-    house: 'hufflepuff',
-    crest:
-      'https://static.wikia.nocookie.net/pottermore/images/5/5e/Hufflepuff_crest.png'
-  },
-  {
-    house: 'ravenclaw',
-    crest:
-      'https://static.wikia.nocookie.net/pottermore/images/4/4f/Ravenclaw_crest.png'
-  }
-];
-
-const renderToDOM = (divId, content) => {
-  const selectedDiv = document.querySelector(divId);
-  selectedDiv.innerHTML = content;
-};
+import renderToDOM from '../utils/renderToDOM';
+import {
+  students,
+  houses,
+  voldysArmy,
+  createId
+} from '../utils/sample_data/data';
 
 // ********** HTML Components  ********** //
 // the basic HMTL structure of app
@@ -110,15 +86,6 @@ const filterBtnRow = () => {
   renderToDOM('#filter-container', domString);
 };
 
-// Create a new ID for the students
-const createId = (array) => {
-  if (array.length) {
-    const idArray = array.map((el) => el.id);
-    return Math.max(...idArray) + 1;
-  }
-  return 0;
-};
-
 // ********** LOGIC  ********** //
 // sorts student to a house and then place them in the students array
 const sortStudent = (e) => {
@@ -153,8 +120,8 @@ const form = () => {
     required
   />
   <label for="floatingInputValue">Name to be sorted</label>
-<button type="submit" class="btn btn-success">Get Sorted!</button>
-</form>`;
+  <button type="submit" class="btn btn-success">Get Sorted!</button>
+  </form>`;
 
   renderToDOM('#form-container', domString);
